Embedd TokenValidator in BotFrameworkAdapter
